requirements.txt
fastapi
uvicorn
google-cloud-firestore
python-dotenv
firebase-admin
pydantic
httpx